Move LocalFile to generic-filehandle/server
