+++ .gitignore
@@ -1,6 +1,5 @@
 .DS_Store
 node_modules
-browser/
 coverage
 dist
 *.log
