--- src/browser.ts
@@ -1,24 +1,0 @@
-import RemoteFile from './remoteFile'
-import BlobFile from './blobFile'
-import { GenericFilehandle, FilehandleOptions } from './filehandle'
-export * from './filehandle'
-
-function fromUrl(source: string, opts: FilehandleOptions = {}): GenericFilehandle {
-  return new RemoteFile(source, opts)
-}
-function open(
-  maybeUrl?: string,
-  _maybePath?: string,
-  maybeFilehandle?: GenericFilehandle,
-  opts: FilehandleOptions = {},
-): GenericFilehandle {
-  if (maybeFilehandle !== undefined) {
-    return maybeFilehandle
-  }
-  if (maybeUrl !== undefined) {
-    return fromUrl(maybeUrl, opts)
-  }
-  throw new Error('no url, path, or filehandle provided, cannot open')
-}
-
-export { open, fromUrl, RemoteFile, BlobFile }
